backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0

mongoengine==0.29.1
pymongo==4.8.0

PyJWT==2.9.0

celery==5.4.0
redis==5.0.8

paho-mqtt==2.1.0

reportlab==4.2.2
openpyxl==3.1.5

python-dateutil==2.9.0.post0
